backend/Newspaper/requirements.txt
fastapi
uvicorn
pymongo
redis
python-dotenv
feedparser
newspaper3k
requests
apscheduler
scikit-learn
lxml_html_clean